Refactored the Map component into the main Home component of the app.
This was done because there are issues with unnecessary rerendering when the Map is in its own component and imported. Rendering performance seems to be improved.

issue #7
